Fixed runtime error with v2m quantity. Added preliminary m2v output

+++ src/PowerModels/SolarModel.java
@@ -69,8 +69,8 @@ public class SolarModel implements IPowerGeneration{
 		return numberOfPanels;
 	}
 
-	public void setNumberOfPanels(int numberOfPanels) {
-		this.numberOfPanels = numberOfPanels;
+	public void setNumberOfPanels(double numberOfPanels) {
+		this.numberOfPanels = (int) numberOfPanels;
 	}
 
 	public Location getLocation() {
